+++ src/Schedulers.jl
@@ -83,16 +83,19 @@ Data in conjunction with a method application's result.
 
 Attributes
 - `changed`: if `false`, the solution has not been changed by the method application
+- `is_local_optimum`: if `true`, the solution is considered a local optimum in respect to
+    the applied method
 - `terminate`: if `true`, a termination condition has been fulfilled
 - `log_info`: customized log info
 """
 mutable struct Result
     changed::Bool
+    is_local_optimum::Bool
     terminate::Bool
     log_info::String
 end
 
-Result() = Result(true, false, "")
+Result() = Result(true, false, false, "")
 
 
 """